fix: add types.ts and delete mocks
