+++ src/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/app/app';
-import { OfferPreviewData } from './mocks/offers';
 import { Provider } from 'react-redux';
 import { store } from './store';
 import { fetchOffersAction, checkAuthAction } from './store/api-actions';
@@ -19,9 +18,7 @@ root.render(
   <React.StrictMode>
     <Provider store={store}>
       <ToastContainer />
-      <App
-        offers={OfferPreviewData}
-      />
+      <App />
     </Provider>
   </React.StrictMode >
 );
