--- src/mocks/favourite.ts
@@ -1,7 +1,0 @@
-import { OfferPreviewData, OfferPreview } from './offers.ts';
-
-export type FavouriteOffer = OfferPreview;
-
-const favouriteData: FavouriteOffer[] = OfferPreviewData;
-
-export { favouriteData };
