+++ src/services/api.ts
@@ -2,11 +2,8 @@ import axios, {AxiosInstance, AxiosRequestConfig, AxiosResponse, AxiosError} fro
 import {StatusCodes} from 'http-status-codes';
 import {getToken} from './token';
 import {toast} from 'react-toastify';
+import { DetailMessageType } from '../types';
 
-type DetailMessageType = {
-  type: string;
-  message: string;
-}
 
 const StatusCodeMapping: Record<number, boolean> = {
   [StatusCodes.BAD_REQUEST]: true,
